chore: extract message to reply preview as component, add voice message button
